requirements.txt
rsa==4.9
Pillow==9.2.0
ttkbootstrap==1.10.1

